+++ src/components/Card/Card.tsx
@@ -1,3 +1,16 @@
-import React from "react";
+import React from 'react';
 
-export {}
+interface CardProps {
+  id: number;
+  distance: number;
+  name: string;
+}
+
+export const Card: React.FC<CardProps> = ({ id, name, distance }) => {
+  return (
+    <button className='market-card'>
+      <p>{name}</p>
+      <p>{distance}</p>
+    </button>
+  );
+};
