--- src/components/Home/Home.css
@@ -1,19 +1,0 @@
-.home {
-  height: 100vh;
-  width: 100vw;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-/* --------- MEDIA QUERIES -----------  */
-
-@media screen and (min-width: 620px) {
-}
-
-@media screen and (min-width: 900px) {
-}
-
-@media screen and (min-width: 1200px) {
-}
